+++ src/main/java/calculator/util/InputValidator.java
@@ -1,0 +1,17 @@
+package calculator.util;
+
+import calculator.exception.inputValidate.InputValidateErrorMessage;
+
+public class InputValidator {
+    public static String validate(String str) {
+        if (str == null || str.isEmpty()) {
+            return "0";
+        }
+
+        if (!str.startsWith("//") && !str.matches("^\\d.*")) {
+            throw new IllegalArgumentException(InputValidateErrorMessage.INVALID_INPUT.getMessage());
+        }
+
+        return str;
+    }
+}
