--- src/test/java/calculator/calculator/CalculatorTest.java
@@ -1,60 +1,0 @@
-package calculator.calculator;
-
-import calculator.Application;
-import camp.nextstep.edu.missionutils.test.NsTest;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-public class CalculatorTest extends NsTest {
-    @Test
-    @DisplayName("정상적인 문자열 입력")
-    void 정상적인_문자열_입력() {
-        // given
-        String input = "3:4,2";
-
-        // when
-        assertSimpleTest(() -> {
-            run(input);
-
-            // then
-            assertThat(output()).contains("결과 : 9");
-        });
-    }
-
-    @Test
-    @DisplayName("빈 문자열 입력 시 0 반환")
-    void 빈_문자열_입력() {
-        // given
-        String input = "";
-
-        // when
-        assertSimpleTest(() -> {
-            run(input);
-
-            // then
-            assertThat(output()).contains("결과 : 0");
-        });
-    }
-
-    @Test
-    @DisplayName("// 또는 숫자로 시작하지 않는다면 예외 처리")
-    void 올바르지_않은_문자열_입력() {
-        // given
-        String input = "?3,4";
-
-        // when & then
-        assertSimpleTest(() ->
-                assertThatThrownBy(() -> runException(input))
-                        .isInstanceOf(IllegalArgumentException.class)
-        );
-    }
-
-    @Override
-    public void runMain() {
-        Application.main(new String[]{});
-    }
-}
